% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{require_package}
\alias{require_package}
\title{Throws an error if the passed package is not available}
\usage{
require_package(package)
}
\arguments{
\item{package}{name of the package to check if it is available}
}
\description{
Throws an error if the passed package is not available
}
