% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{with_postgres_connection}
\alias{with_postgres_connection}
\title{Runs a temporary postgres instande in a docker container}
\usage{
with_postgres_connection(
  image_name,
  code,
  user = NULL,
  password = NULL,
  database = NULL,
  port = 5432
)
}
\arguments{
\item{image_name}{name of the docker image with postgres to use}

\item{code}{the code to execute that uses the postgres connection}

\item{user}{user used for creating the postgres connection}

\item{password}{password used for creating the postgres connection}

\item{database}{database used for creating the postgres connection}

\item{port}{port of the docker container to expose}
}
\description{
Runs a temporary postgres instande in a docker container
}
\details{
In the passed code you can refer to  the \code{con} variable
that stores the postgres connection. See examples
}
\examples{

if (stevedore::docker_available()) {
  with_postgres_connection("postgres:12.10", {
    DBI::dbGetQuery(con, "select version();")
  })
}

}
